requirements.txt
torchsummary
tqdm
setproctitle
pytorch-lightning
pandas
sklearn
matplotlib
netCDF4
pygrib
torchvision
seaborn
xarray
imageio
einops